requirements.txt
six==1.11.0
tensorflow_gpu==1.12.0
tensorrt==7.0.0.11
pycuda==2019.1.2
loguru==0.2.5
tqdm==4.28.1
numpy==1.16.4
opencv_contrib_python==3.4.1.15
matplotlib==2.2.4
Pillow==7.1.2
scikit_learn==0.22.2.post1
PyYAML==5.3.1
